Move max indent logic to plugin
