+++ packages/lexical-rich-text/src/index.ts
@@ -708,12 +708,10 @@ export function registerRichText(editor: LexicalEditor): () => void {
     ),
     editor.registerCommand(
       INDENT_CONTENT_COMMAND,
-      (maxIndent) => {
+      () => {
         return $handleIndentAndOutdent((block) => {
-          const indent = block.getIndent() + 1;
-          block.setIndent(
-            maxIndent == null ? indent : Math.min(indent, maxIndent as number),
-          );
+          const indent = block.getIndent();
+          block.setIndent(indent + 1);
         });
       },
       COMMAND_PRIORITY_EDITOR,
